requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain~=0.1.16
langchain-openai
python-dotenv~=1.0.1
beautifulsoup4
faiss-cpu
requests~=2.31.0
pydantic~=2.7.0
openai~=1.17.0
instructor
transformers
reportlab
sqlalchemy
genanki
faker
supermemo2
ollama
psycopg2-binary
alembic